telegram.com return a geo-located IP address when resolved
Domain telegram.com return a geolocated IP address when resolved. As mentioned in the README file root domains used for baseline tests must not be geolocated because it will result in only resolvers local to the user being returned as valid.

telegram.com@1.1.1.1 (Cloudflare): 
telegram.com.		4800	IN	A	146.75.34.62
telegram.com@8.8.8.8 (Google):
telegram.com.		3784	IN	A	151.101.70.62

You can check if a domain is geolocated using check-host.net to ping from different locations.
Thanks.